fix(connect)：修复第三方登录遗留cookie问题

+++ buyer/pc/themes/b2b2c_v5/store/user.js
@@ -37,6 +37,10 @@ export const mutations = {
     state.user = ''
     Storage.removeItem('user')
     Storage.removeItem('uid')
+    // 主要针对第三方登录留下的数据
+    const domain = document.domain.split('.').slice(1).join('.')
+    Storage.removeItem('user', { domain })
+    Storage.removeItem('uid', { domain })
   },
   /**
    * 设置访问令牌
@@ -54,7 +58,11 @@ export const mutations = {
    * @param state
    */
   [types.REMOVE_ACCESS_TOKEN](state) {
-    Storage.removeItem('access_token')
+    if (process.client) {
+      Storage.removeItem('access_token')
+      // 主要针对第三方登录留下的数据
+      Storage.removeItem('access_token', { domain: document.domain.split('.').slice(1).join('.') })
+    }
   },
   /**
    * 设置刷新令牌
@@ -73,6 +81,8 @@ export const mutations = {
    */
   [types.REMOVE_REFRESH_TOKEN](state) {
     Storage.removeItem('refresh_token')
+    // 主要针对第三方登录留下的数据
+    Storage.removeItem('refresh_token', { domain: document.domain.split('.').slice(1).join('.') })
   }
 }
